requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
gunicorn==21.2.0